LayoutTests/media/video-play-empty-events-expected.txt

Test that play() from EMPTY network state triggers load() and async play event.

EXPECTED (video.networkState == '0') OK
RUN(video.play())
SCRIPT DONE
EVENT(play)
EVENT(waiting)
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
EVENT(playing)
EXPECTED (video.paused == 'false') OK
END OF TEST
